requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-filter==24.3
django-jazzmin==3.0.0
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
